refactor(cli): split helpers into multiple files
