+++ tests/test_main.py
@@ -14,7 +14,7 @@ def test_help_flag() -> None:
 
 def test_includes_commands() -> None:
     result = runner.invoke(cli, ["--help"])
-    commands = ["version", "ask", "sync", "verify", "add", "init"]
+    commands = ["version", "ask", "sync", "verify", "init"]
     for command in commands:
         assert command in result.stdout
